agents/servicenow_agent/requirements.txt
google-adk
python-dotenv
fastapi
uvicorn
requests
aiohttp
mcp